Extracing interfaces outside the ccf/index
