+++ src/lib/ccf/index.test.ts
@@ -1,5 +1,6 @@
 import {describe, expect, jest, test} from '@jest/globals';
-import {CloudCarbonFootprint, Interpolation} from './index';
+import {CloudCarbonFootprint} from './index';
+import {Interpolation} from "../interfaces";
 
 jest.setTimeout(30000);
