+++ src/lib/interfaces/ccf.ts
@@ -1,0 +1,28 @@
+// consumption information for a single instance
+interface IConsumption {
+  idle?: number;
+  tenPercent?: number;
+  fiftyPercent?: number;
+  hundredPercent?: number;
+  minWatts?: number;
+  maxWatts?: number;
+}
+
+// information about a single compute instance
+export interface IComputeInstance {
+  consumption: IConsumption;
+  embodiedEmission?: number;
+  name: string;
+  vCPUs?: number;
+  maxVCPUs?: number;
+}
+
+export enum Interpolation {
+  LINEAR = 'linear',
+  SPLINE = 'spline',
+}
+
+export interface ICcfResult {
+  e: number;
+  m: number;
+}
